refactor: Song 에 Parts 의존성 추가
