--- backend/src/main/java/shook/shook/part/exception/PartsException.java
@@ -1,11 +1,0 @@
-package shook.shook.part.exception;
-
-public class PartsException extends RuntimeException {
-
-    public static class DuplicatePartExistException extends PartsException {
-
-        public DuplicatePartExistException() {
-            super();
-        }
-    }
-}
