+++ backend/src/main/java/shook/shook/song/application/SongService.java
@@ -1,8 +1,12 @@
 package shook.shook.song.application;
 
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import shook.shook.part.domain.Part;
+import shook.shook.song.application.dto.KillingPartResponse;
+import shook.shook.song.application.dto.KillingPartsResponse;
 import shook.shook.song.application.dto.SongRegisterRequest;
 import shook.shook.song.application.dto.SongResponse;
 import shook.shook.song.domain.Song;
@@ -25,7 +29,7 @@ public class SongService {
     public SongResponse findById(final Long songId) {
         final Song song = songRepository.findById(songId)
             .orElseThrow(SongException.SongNotExistException::new);
-        
+
         return SongResponse.from(song);
     }
 
@@ -34,5 +38,23 @@ public class SongService {
             .orElseThrow(SongException.SongNotExistException::new);
 
         return SongResponse.from(song);
+    }
+
+    public KillingPartResponse showTopKillingPart(final Long songId) {
+        final Song song = songRepository.findById(songId)
+            .orElseThrow(SongException.SongNotExistException::new);
+
+        return song.getTopKillingPart()
+            .map(KillingPartResponse::from)
+            .orElseGet(KillingPartResponse::empty);
+    }
+
+    public KillingPartsResponse showKillingParts(final Long songId) {
+        final Song song = songRepository.findById(songId)
+            .orElseThrow(SongException.SongNotExistException::new);
+
+        final List<Part> killingParts = song.getKillingParts();
+
+        return KillingPartsResponse.of(killingParts);
     }
 }
