+++ backend/src/test/java/shook/shook/part/application/PartServiceTest.java
@@ -11,8 +11,6 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import shook.shook.part.application.dto.KillingPartResponse;
-import shook.shook.part.application.dto.KillingPartsResponse;
 import shook.shook.part.application.dto.PartRegisterRequest;
 import shook.shook.part.domain.Part;
 import shook.shook.part.domain.PartLength;
@@ -114,166 +112,6 @@ class PartServiceTest {
             //when
             //then
             assertThatThrownBy(() -> partService.register(request))
-                .isInstanceOf(SongException.SongNotExistException.class);
-        }
-    }
-
-    @DisplayName("노래의 가장 인기있는 킬링파트를 보여준다.")
-    @Nested
-    class ShowTopKillingPart {
-
-        @DisplayName("킬링파트가 존재할 때")
-        @Test
-        void exist() {
-            //given
-            final Part firstPart = Part.forSave(1, PartLength.SHORT, SAVED_SONG);
-            final Part secondPart = Part.forSave(2, PartLength.SHORT, SAVED_SONG);
-            addPart(SAVED_SONG, firstPart);
-            addPart(SAVED_SONG, secondPart);
-
-            final Vote firstVote = Vote.forSave(firstPart);
-            final Vote secondVote = Vote.forSave(secondPart);
-            final Vote thirdVote = Vote.forSave(secondPart);
-            votePart(firstPart, firstVote);
-            votePart(secondPart, secondVote);
-            votePart(secondPart, thirdVote);
-
-            //when
-            saveAndClearEntityManager();
-            final KillingPartResponse response = partService.showTopKillingPart(SAVED_SONG.getId());
-
-            //then
-            assertThat(response.getStart()).isEqualTo(2);
-            assertThat(response.getEnd()).isEqualTo(7);
-        }
-
-        @DisplayName("킬링파트가 없을 때")
-        @Test
-        void notExist() {
-            //given
-            //when
-            final KillingPartResponse response = partService.showTopKillingPart(SAVED_SONG.getId());
-
-            //then
-            assertThat(response).usingRecursiveComparison().isEqualTo(KillingPartResponse.empty());
-        }
-
-        @DisplayName("조회하려는 노래가 없을 때")
-        @Test
-        void songNotExist() {
-            //given
-            //when
-            //then
-            assertThatThrownBy(() -> partService.showTopKillingPart(0L))
-                .isInstanceOf(SongException.SongNotExistException.class);
-        }
-    }
-
-    @DisplayName("노래의 킬링파트들을 보여준다.")
-    @Nested
-    class ShowKillingParts {
-
-        @DisplayName("킬링파트가 충분히 존재할 때")
-        @Test
-        void existEnough() {
-            //given
-            final Part firstPart = Part.forSave(1, PartLength.SHORT, SAVED_SONG);
-            final Part secondPart = Part.forSave(2, PartLength.SHORT, SAVED_SONG);
-            final Part thirdPart = Part.forSave(3, PartLength.SHORT, SAVED_SONG);
-            final Part fourthPart = Part.forSave(4, PartLength.SHORT, SAVED_SONG);
-            addPart(SAVED_SONG, firstPart);
-            addPart(SAVED_SONG, secondPart);
-            addPart(SAVED_SONG, thirdPart);
-            addPart(SAVED_SONG, fourthPart);
-
-            final Vote firstVote = Vote.forSave(firstPart);
-            final Vote secondVote = Vote.forSave(secondPart);
-            final Vote thirdVote = Vote.forSave(secondPart);
-            final Vote fourthVote = Vote.forSave(secondPart);
-            final Vote fifthVote = Vote.forSave(thirdPart);
-            final Vote sixthVote = Vote.forSave(thirdPart);
-            votePart(firstPart, firstVote);
-            votePart(secondPart, secondVote);
-            votePart(secondPart, thirdVote);
-            votePart(secondPart, fourthVote);
-            votePart(thirdPart, fifthVote);
-            votePart(thirdPart, sixthVote);
-
-            //when
-            saveAndClearEntityManager();
-            final KillingPartsResponse response = partService.showKillingParts(SAVED_SONG.getId());
-
-            //then
-            assertThat(response.getResponses()).usingRecursiveComparison()
-                .isEqualTo(List.of(
-                    KillingPartResponse.of(2, 7),
-                    KillingPartResponse.of(3, 8),
-                    KillingPartResponse.of(1, 6)
-                ));
-        }
-
-        @DisplayName("킬링파트가 충분하지 않을 때 ( 2개일 때 )")
-        @Test
-        void exist_notEnough_two() {
-            //given
-            final Part firstPart = Part.forSave(1, PartLength.SHORT, SAVED_SONG);
-            final Part secondPart = Part.forSave(2, PartLength.SHORT, SAVED_SONG);
-            addPart(SAVED_SONG, firstPart);
-            addPart(SAVED_SONG, secondPart);
-
-            final Vote firstVote = Vote.forSave(firstPart);
-            final Vote secondVote = Vote.forSave(secondPart);
-            final Vote thirdVote = Vote.forSave(secondPart);
-            votePart(firstPart, firstVote);
-            votePart(secondPart, secondVote);
-            votePart(secondPart, thirdVote);
-
-            //when
-            saveAndClearEntityManager();
-            final KillingPartsResponse response = partService.showKillingParts(SAVED_SONG.getId());
-
-            //then
-            assertThat(response.getResponses()).usingRecursiveComparison()
-                .isEqualTo(List.of(KillingPartResponse.of(2, 7), KillingPartResponse.of(1, 6)));
-        }
-
-        @DisplayName("킬링파트가 충분하지 않을 때 ( 1개일 때 )")
-        @Test
-        void exist_notEnough_one() {
-            //given
-            final Part firstPart = Part.forSave(1, PartLength.SHORT, SAVED_SONG);
-            addPart(SAVED_SONG, firstPart);
-
-            final Vote vote = Vote.forSave(firstPart);
-            votePart(firstPart, vote);
-
-            //when
-            saveAndClearEntityManager();
-            final KillingPartsResponse response = partService.showKillingParts(SAVED_SONG.getId());
-
-            //then
-            assertThat(response.getResponses()).usingRecursiveComparison()
-                .isEqualTo(List.of(KillingPartResponse.of(1, 6)));
-        }
-
-        @DisplayName("킬링파트가 없을 때")
-        @Test
-        void notExist() {
-            //given
-            //when
-            final KillingPartsResponse response = partService.showKillingParts(SAVED_SONG.getId());
-
-            //then
-            assertThat(response.getResponses()).isEmpty();
-        }
-
-        @DisplayName("조회할 노래가 없을 때")
-        @Test
-        void songNotExist() {
-            //given
-            //when
-            //then
-            assertThatThrownBy(() -> partService.showKillingParts(0L))
                 .isInstanceOf(SongException.SongNotExistException.class);
         }
     }
